added delete feed and comment

+++ src/main/java/com/feed/CommentServlet.java
@@ -5,20 +5,14 @@ import java.io.IOException;
 import java.io.PrintWriter;
 import java.text.ParseException;
 import java.util.Date;
-import java.util.List;
 import java.util.UUID;
-
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-
 import org.json.JSONException;
 import org.json.JSONObject;
-
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.appengine.api.datastore.EntityNotFoundException;
 import com.google.appengine.repackaged.org.joda.time.DateTime;
 
@@ -46,18 +40,30 @@ public class CommentServlet extends HttpServlet {
 			JSONObject comments;
 			try {
 				comments = comment.getSingleComment(c);
-				JSONObject obj=new JSONObject();
-				response.setStatus(200);
-				obj.put("status", "success");
-				obj.put("code","200");
-				obj.put("comment", comments);
-				out.println(obj);
-			} catch (IOException | EntityNotFoundException e) {
+				if (comments.length()>=1) 
+				{
+					response.setStatus(200);
+					comments.put("code", "200");
+					comments.put("success", true);
+					out.println(comments);
+				}
+				else
+				{
+					response.setStatus(400);
+					comments.put("code", "400");
+					JSONObject j=new JSONObject();
+					j.put("success", false);
+					j.put("details","The comment is deleted");
+					comments.put("error",j);
+					out.println(comments);
+				}
+			} 
+			catch (IOException | EntityNotFoundException e) {
 				response.setStatus(500);
 		    	JSONObject obj=new JSONObject();
 		    	JSONObject obj1=new JSONObject();
 				obj1.put("code", "500");
-				obj.put("status", "failed");
+				obj.put("success", false);
 				obj1.put("details", "Enitity not found");
 				obj.put("error", obj1);
 				out.println(obj);
@@ -67,7 +73,7 @@ public class CommentServlet extends HttpServlet {
 		    	JSONObject obj=new JSONObject();
 		    	JSONObject obj1=new JSONObject();
 				obj1.put("code", "500");
-				obj.put("status", "failed");
+				obj.put("success", false);
 				obj1.put("details", "Server error");
 				obj.put("error", obj1);
 				e.printStackTrace();
@@ -81,16 +87,30 @@ public class CommentServlet extends HttpServlet {
 			JSONObject comments;
 			try {
 				comments = comment.getComments(c);
-				response.setStatus(200);
-				comments.put("status", "success");
-				comments.put("code","200");
-				out.println(comments);
+				System.out.println(comments.length());
+				if (comments.length()!=0) 
+				{
+					response.setStatus(200);
+					comments.put("code", "200");
+					comments.put("success", true);
+					out.println(comments);
+				}
+				else
+				{
+					response.setStatus(400);
+					comments.put("code", "400");
+					JSONObject j=new JSONObject();
+					j.put("success", false);
+					j.put("details","The feed is not found or deleted");
+					comments.put("error",j);
+					out.println(comments);
+				}
 			} catch (IOException | EntityNotFoundException e) {
 				response.setStatus(500);
 		    	JSONObject obj=new JSONObject();
 		    	JSONObject obj1=new JSONObject();
 				obj1.put("code", "500");
-				obj.put("status", "failed");
+				obj.put("success", false);
 				obj1.put("details", "Enitity not found");
 				obj.put("error", obj1);
 				out.println(obj);
@@ -100,7 +120,7 @@ public class CommentServlet extends HttpServlet {
 		    	JSONObject obj=new JSONObject();
 		    	JSONObject obj1=new JSONObject();
 				obj1.put("code", "500");
-				obj.put("status", "failed");
+				obj.put("success", false);
 				obj1.put("details", "Server error");
 				obj.put("error", obj1);
 				out.println(obj);
@@ -142,11 +162,12 @@ public void doPost(HttpServletRequest request, HttpServletResponse response)
 				c.setLikes(0);
 				comment.addComment(c);
 			    JSONObject rjson=new JSONObject(c);
+			    rjson.remove("like");
 				response.setStatus(200);
 				JSONObject obj=new JSONObject();
 				obj.put("code", "200");
-				obj.put("status", "success");
-				obj.put("data", rjson);
+				obj.put("success", true);
+				obj.put("comment", rjson);
 				out.println(obj);
 	    	}
 	    	else
@@ -155,7 +176,7 @@ public void doPost(HttpServletRequest request, HttpServletResponse response)
 		    	JSONObject obj=new JSONObject();
 		    	JSONObject obj1=new JSONObject();
 				obj1.put("code", "400");
-				obj.put("status", "failed");
+				obj.put("success", false);
 				obj1.put("details", c.getError());
 				obj.put("error", obj1);
 				out.println(obj);
@@ -166,7 +187,7 @@ public void doPost(HttpServletRequest request, HttpServletResponse response)
 	    	JSONObject obj=new JSONObject();
 	    	JSONObject obj1=new JSONObject();
 			obj1.put("code", "500");
-			obj.put("status", "failed");
+			obj.put("success", false);
 			obj1.put("details", "Enitity not found");
 			obj.put("error", obj1);
 			out.println(obj);
@@ -182,10 +203,7 @@ protected void doPut(HttpServletRequest request, HttpServletResponse response) t
 	Validator validator=new Validator();
 	CommentDao comment=new CommentOperations();
 	PrintWriter out=response.getWriter();
-	String pathInfo = request.getPathInfo(); 
-	String[] pathParts = pathInfo.split("/");
-	String feedId = pathParts[1];
-	String commentId=pathParts[2];
+
 	try 
     {    
 	    StringBuffer jb = new StringBuffer();
@@ -195,68 +213,72 @@ protected void doPut(HttpServletRequest request, HttpServletResponse response) t
 	        jb.append(line);
 	    String str=jb.toString();
 		JSONObject json = new JSONObject(str);
+    	if(validator.isValidCommentUpdate(json,c))
+    	{
+			    if(json.get("like").equals("false")) {
+			    	
+				        DateTime now = new DateTime();
+				        Date millis=new Date(now.getMillis());
+				        c.setLike(false);
+						c.setComment(json.get("comment").toString());
+						c.setFeed_id(json.get("feedId").toString());
+						c.setComment_id(json.get("commentId").toString());
+						c.setDate(millis);
+						comment.updateComment(c);
+						JSONObject rjson=new JSONObject(c);
+						response.setStatus(200);
+						JSONObject obj=new JSONObject();
+						obj.put("code", "200");
+						obj.put("success", true);
+						rjson.remove("like");
+						obj.put("comment", rjson);
+						out.println(obj);
+		
+			    }
+			    
+			    else if(json.get("like").equals("true"))
+			    {
 
-	    if(json.get("like").equals("false")) {
-	    	
-		        DateTime now = new DateTime();
-		        Date millis=new Date(now.getMillis());
-		    	if(validator.isValidCommentUpdate(json,c))
-		    	{
-				    c.setComment(json.get("comment").toString());
-				    c.setFeed_id(json.get("feedId").toString());
-				    c.setComment_id(json.get("commentId").toString());
-				    c.setDate(millis);
-				   
-				    comment.updateComment(c);
-				    JSONObject rjson=new JSONObject(c);
-					response.setStatus(200);
-					JSONObject obj=new JSONObject();
-					obj.put("code", "200");
-					obj.put("status", "success");
-					obj.put("data", rjson);
-					out.println(obj);
-				}
-		    	else 
-		    	{
-			    	response.setStatus(400);
-			    	JSONObject obj=new JSONObject();
-			    	JSONObject obj1=new JSONObject();
-					obj1.put("code", "400");
-					obj.put("status", "failed");
-					obj1.put("details", c.getError());
-					obj.put("error", obj1);
-					out.println(obj);
-		    	}
-	    }
-	    
-	    else if(json.get("like").equals("true"))
-	    {
-				c.setFeed_id(feedId);
-	    		c.setComment_id(commentId);
-			    comment.setLike(c);
-			    int l=comment.getLike(c);
-			    JSONObject obj1=new JSONObject();
-
-			    JSONObject obj=new JSONObject();
-				obj.put("code", "200");
-				obj.put("status", "success");
-				obj1.put("commentId", commentId);
-				obj1.put("likes", l);
-				obj.put("data", obj1);
-				response.setStatus(200);
-			    out.println(obj);	    
-	    }
-	    else
-	    {
-		    	response.setStatus(400);
-		    	JSONObject obj=new JSONObject();
-		    	JSONObject obj1=new JSONObject();
-				obj1.put("code", "400");
-				obj.put("status", "failed");
-				obj1.put("details", "Invalid request");
-				obj.put("error", obj1);
-				out.println(obj);
-	    }
+			    		c.setLike(true);
+				        DateTime now = new DateTime();
+				        Date millis=new Date(now.getMillis());		
+						c.setComment(json.get("comment").toString());
+						c.setFeed_id(json.get("feedId").toString());
+						c.setComment_id(json.get("commentId").toString());
+						c.setDate(millis);
+						comment.updateComment(c);
+					    JSONObject obj1=new JSONObject(c);
+					    obj1.remove("like");
+					    JSONObject obj=new JSONObject();
+						obj.put("code", "200");
+						obj.put("success", true);
+						obj.put("comment", obj1);
+						response.setStatus(200);
+					    out.println(obj);	    
+			    }
+			    else
+			    {
+				    	response.setStatus(400);
+				    	JSONObject obj=new JSONObject();
+				    	JSONObject obj1=new JSONObject();
+						obj1.put("code", "400");
+						obj.put("success", false);
+						obj1.put("details", "Invalid request");
+						obj.put("error", obj1);
+						out.println(obj);
+			    }
+		}
+    	else 
+    	{
+	    	response.setStatus(400);
+	    	JSONObject obj=new JSONObject();
+	    	JSONObject obj1=new JSONObject();
+			obj1.put("code", "400");
+			obj.put("success", false);
+			obj1.put("details", c.getError());
+			obj.put("error", obj1);
+			out.println(obj);
+    	}
 
 	} 
     catch (EntityNotFoundException | JSONException | ParseException e) 
@@ -265,7 +287,7 @@ protected void doPut(HttpServletRequest request, HttpServletResponse response) t
     	JSONObject obj=new JSONObject();
     	JSONObject obj1=new JSONObject();
 		obj1.put("code", "500");
-		obj.put("status", "failed");
+		obj.put("success", false);
 		obj1.put("details", "Enitity not found");
 		obj.put("error", obj1);
 		out.println(obj);
@@ -286,23 +308,47 @@ protected void doDelete(HttpServletRequest request, HttpServletResponse response
 	c.setComment_id(commentId);
 	c.setFeed_id(feedId);
 	try {
-		comment.deleteComment(c);
-		response.setStatus(200);
-	    JSONObject obj=new JSONObject();
-		obj.put("code", "200");
-		obj.put("status", "success");
-		obj.put("commentId", commentId);
-		response.setStatus(200);
-	    out.println(obj);	   
-
+		JSONObject deleted=new JSONObject();
+		deleted=comment.getSingleComment(c);
+		if(deleted.length()>0)
+		{
+			comment.deleteComment(c);
+			response.setStatus(200);
+		    JSONObject obj=new JSONObject();
+			obj.put("code", "200");
+			obj.put("success", true);
+			obj.put("comment", deleted);
+			response.setStatus(200);
+		    out.println(obj);	   
+		}
+		else
+		{
+			response.setStatus(500);
+	    	JSONObject obj=new JSONObject();
+	    	JSONObject obj1=new JSONObject();
+			obj1.put("code", "500");
+			obj.put("success", false);
+			obj1.put("details", "Comment not found or already deleted");
+			obj.put("error", obj1);
+			out.println(obj);
+		}
 	} 
 	catch (EntityNotFoundException e) {
 		response.setStatus(500);
     	JSONObject obj=new JSONObject();
     	JSONObject obj1=new JSONObject();
 		obj1.put("code", "500");
-		obj.put("status", "failed");
+		obj.put("success", false);
 		obj1.put("details", "Enitity not found");
+		obj.put("error", obj1);
+		out.println(obj);
+		e.printStackTrace();
+	} catch (ParseException e) {
+    	JSONObject obj=new JSONObject();
+    	JSONObject obj1=new JSONObject();
+		obj1.put("code", "500");
+		obj.put("success", false);
+		obj1.put("details", "Server Error");
 		obj.put("error", obj1);
 		out.println(obj);
 		e.printStackTrace();
